pyproject.toml
[project]
name = "prj-huggingface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "roboflow>=1.2.16",
    "streamlit>=1.55.0",
    "transformers>=5.3.0",
    "ultralytics>=8.4.22",
]
